+++ src/parser/model/cloudformation-resource.ts
@@ -2,6 +2,7 @@ import md5 = require('md5');
 import { ConsoleUtil } from '../../console-util';
 import { OrgFormationError } from '../../org-formation-error';
 import { IOrganizationBinding, IResource, IResourceRef, TemplateRoot } from '../parser';
+import { OrganizationBindingsSection } from './organization-bindings-section';
 import { Resource } from './resource';
 
 export class CloudFormationResource extends Resource {
@@ -17,18 +18,25 @@ export class CloudFormationResource extends Resource {
     private foreach: IOrganizationBinding;
     private binding: IOrganizationBinding;
 
-    constructor(root: TemplateRoot, id: string, resource: IResource, defaultBinding?: IOrganizationBinding, defaultRegion?: string | string[]) {
+    constructor(root: TemplateRoot, id: string, resource: IResource) {
         super(root, id, resource);
+
+        const bindingsSection = root.bindingSection;
 
         if ((this.resource as any).OrganizationBindings) {
             throw new OrgFormationError(`Resource ${id} has an OrganizationBindings attribute which must be OrganizationBinding (without the S)`);
         }
 
-        this.binding = this.resource.OrganizationBinding as IOrganizationBinding;
+        const bindingExpression =  this.resource.OrganizationBinding;
+        if (bindingExpression && bindingExpression.Ref) {
+            this.binding = bindingsSection.getBinding(bindingExpression.Ref);
+        } else {
+            this.binding = this.resource.OrganizationBinding as IOrganizationBinding;
+        }
 
         if (!this.binding) {
-            this.binding = defaultBinding;
-            this.resource.OrganizationBinding = defaultBinding;
+            this.binding = bindingsSection.defaultBinding;
+            this.resource.OrganizationBinding = bindingsSection.defaultBinding;
         }
 
         if (!this.binding) {
@@ -36,14 +44,19 @@ export class CloudFormationResource extends Resource {
         }
 
         if (!this.binding.Region) {
-            this.binding.Region = defaultRegion;
+            this.binding.Region = bindingsSection.defaultRegion;
         }
 
         super.throwForUnknownAttributes(this.binding, id + '.OrganizationBinding', 'OrganizationalUnit', 'Account', 'ExcludeAccount', 'Region', 'IncludeMasterAccount', 'AccountsWithTag');
 
-        this.foreach = this.resource.Foreach as IOrganizationBinding;
+        const foreachExpression = this.resource.Foreach;
+        if (foreachExpression && foreachExpression.Ref) {
+            this.foreach = bindingsSection.getBinding(foreachExpression.Ref);
+        } else {
+            this.foreach = this.resource.Foreach as IOrganizationBinding;
+        }
         if (this.foreach) {
-            super.throwForUnknownAttributes(this.foreach, id + '.Foreach', 'OrganizationalUnit', 'Account', 'ExcludeAccount', 'IncludeMasterAccount', 'AccountsWithTag');
+            super.throwForUnknownAttributes(this.foreach, id + '.Foreach', 'OrganizationalUnit', 'Account', 'ExcludeAccount',  'Region', 'IncludeMasterAccount', 'AccountsWithTag');
         }
 
         if (this.binding && this.binding.Region) {
@@ -74,10 +87,10 @@ export class CloudFormationResource extends Resource {
 
     public resolveRefs() {
         if (this.binding) {
-            this.normalizedBoundAccounts = this.resolveNormalizedLogicalAccountIds(this.binding);
+            this.normalizedBoundAccounts = this.root.resolveNormalizedLogicalAccountIds(this.binding);
         }
         if (this.foreach) {
-            this.normalizedForeachAccounts = this.resolveNormalizedLogicalAccountIds(this.foreach);
+            this.normalizedForeachAccounts = this.root.resolveNormalizedLogicalAccountIds(this.foreach);
         }
 
         if (this.dependsOnAccountRef) {
@@ -99,63 +112,4 @@ export class CloudFormationResource extends Resource {
         }
     }
 
-    private resolveNormalizedLogicalAccountIds(binding: IOrganizationBinding): string[] {
-        this.throwForAccountIDs(binding.Account);
-        this.throwForAccountIDs(binding.ExcludeAccount);
-
-        const organizationAccountsAndMaster = [this.root.organizationSection.masterAccount, ...this.root.organizationSection.accounts];
-        const accounts = super.resolve(binding.Account, binding.Account === '*' ? this.root.organizationSection.accounts : organizationAccountsAndMaster);
-        const excludeAccounts = super.resolve(binding.ExcludeAccount, binding.ExcludeAccount === '*' ? this.root.organizationSection.accounts : organizationAccountsAndMaster);
-        const organizationalUnits = super.resolve(binding.OrganizationalUnit, this.root.organizationSection.organizationalUnits);
-
-        const accountLogicalIds = accounts.map((x) => x.TemplateResource!.logicalId);
-        const result = new Set<string>(accountLogicalIds);
-        for (const unit of organizationalUnits) {
-            const accountsForUnit = unit.TemplateResource!.accounts.map((x) => x.TemplateResource!.logicalId);
-            for (const logicalId of accountsForUnit) {
-                result.add(logicalId);
-            }
-        }
-        if (binding.IncludeMasterAccount) {
-            if (this.root.organizationSection.masterAccount) {
-                result.add(this.root.organizationSection.masterAccount.logicalId);
-            } else {
-                new OrgFormationError('unable to include master account if master account is not part of the template');
-            }
-        }
-
-        if (binding.AccountsWithTag) {
-            const tagToMatch = binding.AccountsWithTag;
-            const accountsWithTag = this.root.organizationSection.findAccounts((x) => (x.tags !== undefined) && Object.keys(x.tags).indexOf(tagToMatch) !== -1);
-            for (const account of accountsWithTag.map((x) => x.logicalId)) {
-                result.add(account);
-            }
-        }
-
-        for (const account of excludeAccounts.map((x) => x.TemplateResource!.logicalId)) {
-            result.delete(account);
-        }
-
-        return [...result];
-    }
-
-    private throwForAccountIDs(resourceRefs: IResourceRef | IResourceRef[]) {
-
-        if (resourceRefs) {
-            if (typeof resourceRefs === 'string') {
-                if (resourceRefs.match(/\d{12}/)) {
-                    throw new OrgFormationError(`error with account binding on ${resourceRefs}. Directly binding on accountid is not supported, use !Ref logicalId instead.`);
-                }
-                if (Array.isArray(resourceRefs)) {
-                    for (const elm of resourceRefs) {
-                        if (typeof elm === 'string') {
-                            if (elm.match(/\d{12}/)) {
-                                throw new OrgFormationError(`error with account binding on ${elm}. Directly binding on accountid is not supported, use !Ref logicalId instead.`);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
 }