+++ test/unit-tests/cfn-binder/cfn-template.test.ts
@@ -142,7 +142,7 @@ describe('when creating cloudformation with output section', () => {
         };
 
         templateRoot.resourcesSection.resources.push(target.resources[0]);
-        templateRoot.resourcesSection.resources.push(new CloudFormationResource(templateRoot, 'bucket2', {Type: 'abcdef'}, templateRoot.contents.OrganizationBinding));
+        templateRoot.resourcesSection.resources.push(new CloudFormationResource(templateRoot, 'bucket2', {Type: 'abcdef'}));
 
         templateRoot.contents.Outputs = {
             Output : {